+++ lib/online_migrations.rb
@@ -83,14 +83,22 @@ module OnlineMigrations
     end
 
     # Run background data migrations
-    def run_background_migrations
-      BackgroundMigrations::Scheduler.run
+    #
+    # @option options [String, Symbol, nil] :shard The name of the shard to run
+    #   background data migrations on. By default runs on all shards.
+    #
+    def run_background_migrations(**options)
+      BackgroundMigrations::Scheduler.run(**options)
     end
     alias run_background_data_migrations run_background_migrations
 
     # Run background schema migrations
-    def run_background_schema_migrations
-      BackgroundSchemaMigrations::Scheduler.run
+    #
+    # @option options [String, Symbol, nil] :shard The name of the shard to run
+    #   background schema migrations on. By default runs on all shards.
+    #
+    def run_background_schema_migrations(**options)
+      BackgroundSchemaMigrations::Scheduler.run(**options)
     end
 
     def deprecator