+++ lib/online_migrations/background_migrations/scheduler.rb
@@ -9,20 +9,27 @@ module OnlineMigrations
     # successive runs has passed.
     #
     # Scheduler should be configured to run periodically, for example, via cron.
+    #
     # @example Run via whenever
     #   # add this to schedule.rb
     #   every 1.minute do
-    #     runner "OnlineMigrations.run_background_migrations"
+    #     runner "OnlineMigrations.run_background_data_migrations"
+    #   end
+    #
+    # @example Run via whenever (specific shard)
+    #   every 1.minute do
+    #     runner "OnlineMigrations.run_background_data_migrations(shard: :shard_two)"
     #   end
     #
     class Scheduler
-      def self.run
-        new.run
+      def self.run(**options)
+        new.run(**options)
       end
 
       # Runs Scheduler
-      def run
+      def run(**options)
         active_migrations = Migration.runnable.active.queue_order
+        active_migrations = active_migrations.where(shard: options[:shard]) if options.key?(:shard)
         runnable_migration = active_migrations.select(&:interval_elapsed?).first
 
         if runnable_migration