+++ test/background_migrations/scheduler_test.rb
@@ -16,6 +16,7 @@ module BackgroundMigrations
     def teardown
       @connection.drop_table(:users, if_exists: true)
       OnlineMigrations::BackgroundMigrations::Migration.delete_all
+      on_each_shard { Dog.delete_all }
     end
 
     def test_run
@@ -23,7 +24,7 @@ module BackgroundMigrations
       user2 = User.create!
       user3 = User.create!
 
-      m = OnlineMigrations::BackgroundMigrations::Migration.create!(
+      m = create_migration(
         migration_name: "MakeAllNonAdmins",
         batch_size: 2,
         sub_batch_size: 1,
@@ -49,11 +50,29 @@ module BackgroundMigrations
       end
     end
 
+    def test_run_specific_shard
+      on_each_shard { Dog.create! }
+
+      m = create_migration(migration_name: "MakeAllDogsNice")
+
+      scheduler = OnlineMigrations::BackgroundMigrations::Scheduler.new
+      scheduler.run(shard: :shard_two)
+      scheduler.run(shard: :shard_two) # finish
+
+      assert m.reload.running?
+
+      shard_one_migration = m.children.find_by(shard: :shard_one)
+      assert shard_one_migration.enqueued?
+
+      shard_two_migration = m.children.find_by(shard: :shard_two)
+      assert shard_two_migration.succeeded?
+    end
+
     def test_run_migration_has_stuck_job
       user1 = User.create!
       user2 = User.create!
 
-      m = OnlineMigrations::BackgroundMigrations::Migration.create!(
+      m = create_migration(
         migration_name: "MakeAllNonAdmins",
         batch_size: 1,
         sub_batch_size: 1
@@ -81,5 +100,10 @@ module BackgroundMigrations
       scheduler.run # last run to ensure there are no more work
       assert m.reload.completed?
     end
+
+    private
+      def create_migration(migration_name:, **attributes)
+        @connection.create_background_data_migration(migration_name, **attributes)
+      end
   end
 end